// frontend/src/types.ts
export interface OptionType {
  readonly value: string;
  readonly label: string;
}
// ...
export interface TupleType {
  value: string;
  label: string;
  source: string;
  target: string;
  type: string;
  sourceType?: string;
}